requirements.txt
# Dev
black
flake8
ipython
mypy

# Prod
jupyterlab
pandas
seaborn
speedtest-cli